Updated DialogText to be a AnimatedRichText.
Improved how the DialogText renders above the speaking character.
Some interface elements hid when a character moves.
Updated the Character node to be a Node2D and to have a Sprite and an AnimationPlayer.
Installed AsepriteImporter plugin to create character animations from an
Aseprite PNG + JSON.
Created the DialogBox node to render "narrator" texts.