Created nodes to handle the Inventory.
Added the inventory container to the Interface layer.
Created script templates for inventory items and room props.
Inventory items must be registered in the Inventory in order to make
them pickable for the player.
Now props can have custom behaviors using the PropTemplate.
Added the bucket to the Room.tscn for testing adding the Bucket to the
inventory.
Created the Prop node and the Prop scene.